fix: address CodeRabbit review on ESM-C/ESMFold2 oracles
- Critical: make EmbeddingResult.save_attributes an instance method (was a
  @classmethod referencing cls.embeddings, which fails when callbacks call it on
  the result instance). Fixed for both ESMCResult and ESM2Result.
- Minor: default local_plddt in ESMFold2._reduce_output is now 2D (1, N) to
  match ptm/pae and keep save_attributes working.
- Cleanup: extract the duplicated validate_array_range into folding/utils.py
  (used by esmfold and esmfold2), and factor the shared mock_fold body in the
  esmfold/esmfold2 test fixtures into a _build_mock_fold_result helper.

Full unit suite green (112 passed, 4 skipped); ruff clean.
