+++ src/bagel/oracles/embedding/esmc.py
@@ -30,9 +30,8 @@ class ESMCResult(EmbeddingResult):
     input_chains: list[Chain]
     embeddings: npt.NDArray[np.float64]
 
-    @classmethod
-    def save_attributes(cls, filepath: pl.Path) -> None:
-        np.savetxt(filepath.with_suffix('.embeddings'), cls.embeddings, fmt='%.6f', header='embeddings')
+    def save_attributes(self, filepath: pl.Path) -> None:
+        np.savetxt(filepath.with_suffix('.embeddings'), self.embeddings, fmt='%.6f', header='embeddings')
 
 
 class ESMC(EmbeddingOracle):
